requirements.txt
alabaster==0.7.10
argon2-cffi==16.3.0
awesome-slugify==1.6.5
Babel==2.4.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
coverage==4.4.1
decorator==4.0.11
defusedxml==0.5.0
Django==1.10.7
django-allauth==0.32.0
django-appconf==1.0.2
django-braces==1.11.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.3
django-extensions==1.7.9
django-model-utils==3.0.0
django-redis==4.8.0
django-test-plus==1.0.17
docutils==0.13.1
factory-boy==2.8.1
Faker==0.7.17
idna==2.5
imagesize==0.7.1
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.9.6
jsonfield==1.0.3
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pinax-stripe==3.4.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.34
pycparser==2.17
Pygments==2.2.0
pytest==3.1.2
pytest-django==3.1.2
pytest-sugar==0.8.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
regex==2017.6.7
requests==2.18.1
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.2
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
stripe==1.59.0
termcolor==1.1.0
traitlets==4.3.2
Unidecode==0.4.20
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.12.2
whitenoise==3.3.0
